Improve source documentation
