+++ internal/config/config.go
@@ -7,23 +7,25 @@ import (
 	"github.com/joho/godotenv"
 )
 
+// Config stores the app configuration.
 type Config struct {
-	OpenaiToken string
-	DeeplToken  string
+	OpenAIToken string
+	DeepLToken  string
 	Addr        string
 	TlsCert     string
 	TlsKey      string
 	DSN         string
 }
 
+// Load loads Config, using .env as the config source, and returns it.
 func Load() (*Config, error) {
 	if err := godotenv.Load(".env"); err != nil {
 		return nil, fmt.Errorf("failed to load .env: %w", err)
 	}
 
 	return &Config{
-		OpenaiToken: os.Getenv("OPENAI_TOKEN"),
-		DeeplToken:  os.Getenv("DEEPL_TOKEN"),
+		OpenAIToken: os.Getenv("OPENAI_TOKEN"),
+		DeepLToken:  os.Getenv("DEEPL_TOKEN"),
 		Addr:        os.Getenv("SERVER_ADDR"),
 		TlsCert:     os.Getenv("TLS_CERT"),
 		TlsKey:      os.Getenv("TLS_KEY"),
