+++ internal/models/models.go
@@ -2,6 +2,9 @@ package models
 
 import "fmt"
 
+// TODO: This file is wip
+
+// Word represents internal word structure
 type Word struct {
 	Phrase     string `json:"phrase"`
 	Definition string `json:"definition"`
